Fix pickling in Python 2 and 3
Explicitly use object or dict methods for getting attributes. This keeps
the `__slots__` as object attributes and everything else as part of the
dict instance.

Pickle can then automatically take care of the dict and the only
relevant slot (`__keys_order__`) is saved. No need to update a
dictionary with the items it already has.