Make `RepeatUntil` take in-progress `Container`
This makes it easy to implement things that depend on history. For example
parsing fields that have a maximum length or a terminal condition:

RepeatUntil(lambda obj, ctx: (ctx[-1] == 4) or (len(ctx)>= 43), Byte)

This would parse bytes until one of them was 4 or 43 of them were read,
whichever happens first.